+++ libbeat/cmd/instance/beat_test.go
@@ -63,7 +63,6 @@ func TestNewInstance(t *testing.T) {
 	}
 	assert.Equal(t, "testbeat", b.Info.Beat)
 	assert.Equal(t, "testbeat", b.Info.IndexPrefix)
-
 }
 
 func TestNewInstanceUUID(t *testing.T) {
@@ -140,7 +139,7 @@ func TestEmptyMetaJson(t *testing.T) {
 	defer os.Remove(metaPath)
 
 	// load metadata
-	err = b.loadMeta(metaPath)
+	err = b.LoadMeta(metaPath)
 
 	assert.Equal(t, nil, err, "Unable to load meta file properly")
 	assert.NotEqual(t, uuid.Nil, b.Info.ID, "Beats UUID is not set")
@@ -162,7 +161,7 @@ func TestMetaJsonWithTimestamp(t *testing.T) {
 	metaFile.Close()
 	defer os.Remove(metaPath)
 
-	err = firstBeat.loadMeta(metaPath)
+	err = firstBeat.LoadMeta(metaPath)
 	assert.Equal(t, nil, err, "Unable to load meta file properly")
 
 	secondBeat, err := NewBeat("filebeat", "testidx", "0.9", false, nil)
@@ -171,7 +170,7 @@ func TestMetaJsonWithTimestamp(t *testing.T) {
 	}
 	secondBeat.Info.Logger = logger
 	assert.False(t, firstStart.Equal(secondBeat.Info.FirstStart), "Before meta.json is loaded, first start must be different")
-	err = secondBeat.loadMeta(metaPath)
+	err = secondBeat.LoadMeta(metaPath)
 	require.NoError(t, err)
 
 	assert.Equal(t, nil, err, "Unable to load meta file properly")
@@ -258,7 +257,7 @@ elasticsearch:
 
 		update := &reload.ConfigWithMeta{Config: c}
 		m := &outputReloaderMock{}
-		reloader := b.makeOutputReloader(m)
+		reloader := b.MakeOutputReloader(m)
 
 		require.False(t, b.Config.Output.IsSet(), "the output should not be set yet")
 		require.True(t, b.isConnectionToOlderVersionAllowed(), "allow_older_versions flag should be true from 8.11")
@@ -346,7 +345,7 @@ output:
 				},
 			}}
 
-			err = promoteOutputQueueSettings(b)
+			err = PromoteOutputQueueSettings(b)
 			require.NoError(t, err)
 
 			ms, err := memqueue.SettingsForUserConfig(b.Config.Pipeline.Queue.Config())
