+++ x-pack/filebeat/fbreceiver/factory.go
@@ -10,16 +10,12 @@ import (
 
 	"github.com/elastic/beats/v7/filebeat/beater"
 	"github.com/elastic/beats/v7/filebeat/cmd"
-	"github.com/elastic/beats/v7/libbeat/api"
-	"github.com/elastic/beats/v7/libbeat/cmd/instance"
-	"github.com/elastic/beats/v7/libbeat/otelbeat/beatreceiver"
 	"github.com/elastic/beats/v7/libbeat/processors"
 	"github.com/elastic/beats/v7/libbeat/publisher/processing"
-	"github.com/elastic/beats/v7/libbeat/version"
 	"github.com/elastic/beats/v7/x-pack/filebeat/include"
 	inputs "github.com/elastic/beats/v7/x-pack/filebeat/input/default-inputs"
+	xpInstance "github.com/elastic/beats/v7/x-pack/libbeat/cmd/instance"
 	"github.com/elastic/elastic-agent-libs/mapstr"
-	metricreport "github.com/elastic/elastic-agent-system-metrics/report"
 
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/consumer"
@@ -49,60 +45,18 @@ func createReceiver(_ context.Context, set receiver.Settings, baseCfg component.
 	settings.ElasticLicensed = true
 	settings.Initialize = append(settings.Initialize, include.InitializeModule)
 
-	b, err := instance.NewBeatReceiver(settings, cfg.Beatconfig, true, consumer, set.Logger.Core())
+	b, err := xpInstance.NewBeatForReceiver(settings, cfg.Beatconfig, true, consumer, set.Logger.Core())
 	if err != nil {
 		return nil, fmt.Errorf("error creating %s: %w", Name, err)
 	}
 
 	beatCreator := beater.New(inputs.Init)
-
-	beatConfig, err := b.BeatConfig()
+	br, err := xpInstance.NewBeatReceiver(b, beatCreator, set.Logger)
 	if err != nil {
-		return nil, fmt.Errorf("error getting beat config: %w", err)
+		return nil, fmt.Errorf("error creating %s:%w", Name, err)
 	}
 
-	b.RegisterMetrics()
-
-	statsReg := b.Info.Monitoring.StatsRegistry
-
-	// stats.beat
-	processReg := statsReg.GetRegistry("beat")
-	if processReg == nil {
-		processReg = statsReg.NewRegistry("beat")
-	}
-
-	// stats.system
-	systemReg := statsReg.GetRegistry("system")
-	if systemReg == nil {
-		systemReg = statsReg.NewRegistry("system")
-	}
-
-	err = metricreport.SetupMetrics(b.Info.Logger.Named("metrics"), b.Info.Beat, version.GetDefaultVersion(), metricreport.WithProcessRegistry(processReg), metricreport.WithSystemRegistry(systemReg))
-	if err != nil {
-		return nil, fmt.Errorf("error setting up metrics report: %w", err)
-	}
-
-	if b.Config.HTTP.Enabled() {
-		var err error
-		b.API, err = api.NewWithDefaultRoutes(b.Info.Logger.Named("metrics.http"), b.Config.HTTP, api.NamespaceLookupFunc())
-		if err != nil {
-			return nil, fmt.Errorf("could not start the HTTP server for the API: %w", err)
-		}
-		b.API.Start()
-	}
-
-	fbBeater, err := beatCreator(&b.Beat, beatConfig)
-	if err != nil {
-		return nil, fmt.Errorf("error getting %s creator:%w", Name, err)
-	}
-
-	base := beatreceiver.BeatReceiver{
-		Beat:   b,
-		Beater: fbBeater,
-		Logger: set.Logger,
-	}
-
-	return &filebeatReceiver{BeatReceiver: base}, nil
+	return &filebeatReceiver{BeatReceiver: br}, nil
 }
 
 // copied from filebeat cmd.
