+++ x-pack/libbeat/outputs/otelconsumer/otelconsumer.go
@@ -1,19 +1,6 @@
-// Licensed to Elasticsearch B.V. under one or more contributor
-// license agreements. See the NOTICE file distributed with
-// this work for additional information regarding copyright
-// ownership. Elasticsearch B.V. licenses this file to you under
-// the Apache License, Version 2.0 (the "License"); you may
-// not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
+// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+// or more contributor license agreements. Licensed under the Elastic License;
+// you may not use this file except in compliance with the Elastic License.
 
 package otelconsumer
 
@@ -146,7 +133,7 @@ func (out *otelConsumer) logsPublish(ctx context.Context, batch publisher.Batch)
 		// if data_stream field is set on beatEvent. Add it to logrecord.Attributes to support dynamic indexing
 		if val, _ := beatEvent.GetValue("data_stream"); val != nil {
 			// If the below sub fields do not exist, it will return empty string.
-			var subFields = []string{"dataset", "namespace", "type"}
+			subFields := []string{"dataset", "namespace", "type"}
 
 			for _, subField := range subFields {
 				// value, ok := data.Map().Get(subField)
